+++ app/Auth/layout.tsx
@@ -1,0 +1,17 @@
+import { Box } from "@mui/material";
+import { Outlet } from "react-router";
+export const Layout = () => {
+  return (
+    <Box
+      sx={{
+        minHeight: "100vh",
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <Outlet />
+    </Box>
+  );
+};
