+++ app/routes.ts
@@ -1,6 +1,4 @@
 import { type RouteConfig, index, route } from "@react-router/dev/routes";
+import { flatRoutes } from "@react-router/fs-routes";
 
-export default [
-  index("routes/home.tsx"),
-  route("signin", "routes/signin.tsx"),
-] satisfies RouteConfig;
+export default [...(await flatRoutes())] satisfies RouteConfig;
