+++ app/routes/[.well-known].$.tsx
@@ -1,0 +1,8 @@
+export function loader() {
+  return Response.json({
+    workspace: {
+      root: "D:\\Code\\Auth\\app",
+      uuid: "5bf853b5-a39e-4431-b0b5-75f4d931f9ba",
+    },
+  });
+}
