+++ app/routes/_auth.login.tsx
@@ -1,0 +1,5 @@
+import { Login } from "../Auth";
+
+export default function LoginPage() {
+  return <Login />;
+}
