+++ app/routes/_auth.tsx
@@ -1,0 +1,5 @@
+import { Layout } from "~/Auth/layout";
+
+export default function AuthLayout() {
+  return <Layout />;
+}
