+++ app/routes/signin.tsx
@@ -1,8 +1,6 @@
 import { SocialProviders } from "lib/auth";
 import type { Route } from "./+types/signin";
 
-import { SignIn } from "~/sign-up";
-
 export const loader = async (args: Route.LoaderArgs) => {
   return {
     socialProviders: SocialProviders,
@@ -10,5 +8,5 @@ export const loader = async (args: Route.LoaderArgs) => {
 };
 
 export default function SignInPage(args: Route.ComponentProps) {
-  return <SignIn />;
+  return <div></div>;
 }
