+++ app/routes/signup.tsx
@@ -1,0 +1,5 @@
+import { SignUp } from "~/sign-up";
+
+export default function SignInPage() {
+  return <SignUp />;
+}
