+++ lib/auth-client.ts
@@ -2,17 +2,14 @@ import { createAuthClient } from "better-auth/client";
 import {
   apiKeyClient,
   jwtClient,
-  passkeyClient,
   twoFactorClient,
-  usernameClient,
 } from "better-auth/client/plugins";
+import { passkeyClient } from "@better-auth/passkey/client";
 import { checkNameOrEmailClient } from "./Plugin/checkNameOrEmailClient";
 export const authClient = createAuthClient({
   plugins: [
     twoFactorClient(),
-    usernameClient(),
     passkeyClient(),
-
     apiKeyClient(),
     jwtClient(),
     checkNameOrEmailClient(),
