+++ package.json
@@ -2,6 +2,7 @@
   "name": "auth",
   "private": true,
   "type": "module",
+  "version": "0.0.1",
   "scripts": {
     "build": "react-router build",
     "dev": "cross-env NODE_ENV=development node server.js",
@@ -9,42 +10,44 @@
     "typecheck": "react-router typegen && tsc -b"
   },
   "dependencies": {
+    "@better-auth/passkey": "^1.4.5",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
-    "@mui/icons-material": "^7.3.5",
-    "@mui/material": "^7.3.5",
-    "@prisma/adapter-pg": "^7.0.0",
-    "@prisma/client": "^7.0.0",
-    "@react-router/node": "^7.9.2",
-    "@react-router/serve": "^7.9.2",
-    "better-auth": "^1.3.34",
+    "@mui/icons-material": "^7.3.6",
+    "@mui/material": "^7.3.6",
+    "@prisma/adapter-pg": "^7.1.0",
+    "@prisma/client": "^7.1.0",
+    "@react-router/fs-routes": "^7.10.1",
+    "@react-router/node": "^7.10.1",
+    "@react-router/serve": "^7.10.1",
+    "better-auth": "^1.4.5",
     "compression": "^1.8.1",
     "dotenv": "^17.2.3",
-    "express": "^5.1.0",
-    "isbot": "^5.1.31",
+    "express": "^5.2.1",
+    "isbot": "^5.1.32",
     "morgan": "^1.10.1",
     "pg": "^8.16.3",
-    "react": "^19.1.1",
-    "react-dom": "^19.1.1",
-    "react-router": "^7.10.0",
+    "react": "^19.2.1",
+    "react-dom": "^19.2.1",
+    "react-router": "^7.10.1",
     "react-transition-group": "^4.4.5",
-    "tsx": "^4.20.6",
+    "tsx": "^4.21.0",
     "zod": "^4.1.13"
   },
   "devDependencies": {
-    "@react-router/dev": "^7.10.0",
-    "@tailwindcss/vite": "^4.1.13",
-    "@types/node": "^24",
-    "@types/react": "^19.2.7",
-    "@types/express": "^5.0.5",
+    "@react-router/dev": "^7.10.1",
+    "@tailwindcss/vite": "^4.1.17",
+    "@types/express": "^5.0.6",
     "@types/morgan": "^1.9.10",
+    "@types/node": "^24",
     "@types/pg": "^8.15.6",
-    "@types/react-dom": "^19.1.9",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "cross-env": "^10.1.0",
-    "prisma": "^7.0.0",
-    "tailwindcss": "^4.1.13",
-    "vite": "^7.2.6",
+    "prisma": "^7.1.0",
+    "tailwindcss": "^4.1.17",
     "typescript": "^5.9.3",
+    "vite": "^7.2.6",
     "vite-tsconfig-paths": "^5.1.4"
   }
 }
